update functions.py for python 3 and add an example of a decorator
git-svn-id: https://software.sandia.gov/svn/pyomo/pyomo.data/trunk@9951 570ccb8d-5833-0410-9ce8-cbedd0da42eb

+++ pyomo/data/pyomobook/python/functions.py
@@ -8,9 +8,9 @@ def Apply(f, a):
     return r
  
 def SqifOdd(x):
-    # if x is odd, 2*(x/2) is not x
+    # if x is odd, 2*trunc(x/2) is not x
     # due to integer divide of x/2
-    if 2*(x/2) == x:
+    if 2*int(x/2) == x:
         return x
     else:
         return x*x
